# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Ratinam Crackers Business Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /v1/auth/pin-login:
    post:
      operationId: pinLogin
      tags: [auth]
      summary: Login with PIN (POS)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PinLoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
...
  /v1/auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

  # ============================================================
  # DASHBOARD
  # ============================================================
...
  /v1/agents:
    get:
      operationId: listAgents
      tags: [agents]
      summary: List agents
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Agent list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentListResponse"
    post:
      operationId: createAgent
      tags: [agents]
      summary: Create agent
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgentBody"
      responses:
        "201":
          description: Agent created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
  /v1/agents/{id}:
    get:
      operationId: getAgent
      tags: [agents]
      summary: Get agent by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
    put:
      operationId: updateAgent
      tags: [agents]
      summary: Update agent
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAgentBody"
      responses:
        "200":
          description: Updated agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
  /v1/agents/{id}/commission:
    get:
      operationId: getAgentCommission
      tags: [agents]
      summary: Get agent commission report
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Commission report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommissionResponse"

  # ============================================================
  # ESTIMATES
  # ============================================================
...
  /v1/coupons:
    get:
      operationId: listCoupons
      tags: [coupons]
      summary: List coupons
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Coupon list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CouponListResponse"
    post:
      operationId: createCoupon
      tags: [coupons]
      summary: Create coupon
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCouponBody"
      responses:
        "201":
          description: Coupon created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
  /v1/coupons/public:
    get:
      operationId: listPublicCoupons
      tags: [coupons]
      summary: List active public coupons (website)
      parameters:
        - name: autoApply
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Public coupons
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CouponListResponse"
  /v1/coupons/validate:
    post:
      operationId: validateCoupon
      tags: [coupons]
      summary: Validate a coupon code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateCouponBody"
      responses:
        "200":
          description: Validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CouponValidationResponse"
  /v1/coupons/{id}:
    get:
      operationId: getCoupon
      tags: [coupons]
      summary: Get coupon by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Coupon
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
    put:
      operationId: updateCoupon
      tags: [coupons]
      summary: Update coupon
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCouponBody"
      responses:
        "200":
          description: Updated coupon
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"

  # ============================================================
  # POS
  # ============================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiError:
      type: object
      properties:
        success:
          type: boolean
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer

    # AUTH SCHEMAS
    LoginBody:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    PinLoginBody:
      type: object
      required: [username, pin, locationId]
      properties:
        username:
          type: string
        pin:
          type: string
        locationId:
          type: string
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            accessToken:
              type: string
            user:
              $ref: "#/components/schemas/UserProfile"
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        username:
          type: string
        role:
          type: string
        locationIds:
          type: array
          items:
            type: string
        maxDiscountPct:
          type: number

    # USER SCHEMAS
...
    CommissionTier:
      type: object
      properties:
        from:
          type: number
        to:
          type: number
        rate:
          type: number
    Agent:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        promoCode:
          type: string
        commissionTiers:
          type: array
          items:
            $ref: "#/components/schemas/CommissionTier"
        maxDiscountPct:
          type: number
        monthlyTarget:
          type: number
        status:
          type: string
    AgentListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/Agent"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
    CreateAgentBody:
      type: object
      required: [name, phone]
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        promoCode:
          type: string
        commissionTiers:
          type: array
          items:
            $ref: "#/components/schemas/CommissionTier"
        maxDiscountPct:
          type: number
        monthlyTarget:
          type: number
    UpdateAgentBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        promoCode:
          type: string
        commissionTiers:
          type: array
          items:
            $ref: "#/components/schemas/CommissionTier"
        maxDiscountPct:
          type: number
        monthlyTarget:
          type: number
    CommissionResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            items:
              type: array
              items:
                type: object
                properties:
                  invoiceNo:
                    type: string
                  amount:
                    type: number
                  commission:
                    type: number
                  date:
                    type: string
            totalCommission:
              type: number
            totalSales:
              type: number

    # ESTIMATE SCHEMAS
...
    Coupon:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        type:
          type: string
          enum: [flat, percent, minorder, bxgy, bundle, firstorder, agentpromo]
        discountValue:
          type: number
        maxDiscountCap:
          type: number
        minOrderValue:
          type: number
        usageLimit:
          type: integer
        usedCount:
          type: integer
        perCustomerLimit:
          type: integer
        applicableChannels:
          type: array
          items:
            type: string
        applicableCustomerTypes:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, paused, expired]
        autoApply:
          type: boolean
        validFrom:
          type: string
        validUntil:
          type: string
        description:
          type: string
    CouponListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/Coupon"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
    CreateCouponBody:
      type: object
      required: [code, type, discountValue, validFrom, validUntil]
      properties:
        code:
          type: string
        type:
          type: string
        discountValue:
          type: number
        maxDiscountCap:
          type: number
        minOrderValue:
          type: number
        usageLimit:
          type: integer
        perCustomerLimit:
          type: integer
        applicableChannels:
          type: array
          items:
            type: string
        applicableCustomerTypes:
          type: array
          items:
            type: string
        autoApply:
          type: boolean
        validFrom:
          type: string
        validUntil:
          type: string
        description:
          type: string
    UpdateCouponBody:
      type: object
      properties:
        status:
          type: string
        discountValue:
          type: number
        validUntil:
          type: string
        description:
          type: string
    ValidateCouponBody:
      type: object
      required: [code, cartTotal]
      properties:
        code:
          type: string
        cartTotal:
          type: number
        customerId:
          type: string
        channel:
          type: string
        items:
          type: array
          items:
            type: object
    CouponValidationResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            valid:
              type: boolean
            discountAmount:
              type: number
            discountDescription:
              type: string
            couponId:
              type: string
            error:
              type: string

    # POS SCHEMAS
...